Add csv-to-json trx with gemini
We integrate with Gemini 2.0-Flash via LangChain to convert the source CSV file to the corresponding JSON format. We set the expectation and the validation constraints, and offload such operations usually handled by the local applications to the LLM. We update the environment variable of the Gemini's API key to align with what LangChain expects for the provider. We handle errors and respond to the client appropriately with a formatter error json.